feat: Added new provider google_slides
We were using google_cloud_storage as provider for generating
certificates. But this has an issue, we cache the token. So if say
Bhasini call is ran first then the cache for gcs token will only have single
scope, so next time when we generate certificate, we get a token from cache
which had only single scope, instead of required scopes we need for generating
certificate.