Custom certificate: webhook function to generate certificate  (#4056)
* add the create certificate support

* chore: crud apis wip

* checkout

* test: create-cert failure cases done

* feat: Added all CRUD apis and tests

* refactor: Added cert template type as enum

* refactor: removed dev logs

* fix: removed public validation, will add scope validation

* upload image in gcs

* test cases

* optimization

* fix test case

* resolved reviews

* build: Run CIs for any PRs

* feat: Added new provider google_slides

We were using google_cloud_storage as provider for generating
certificates. But this has an issue, we cache the token. So if say
Bhasini call is ran first then the cache for gcs token will only have single
scope, so next time when we generate certificate, we get a token from cache
which had only single scope, instead of required scopes we need for generating
certificate.

* fix: Removed making copied template public

Since service account is copying the template, the service account will
have access to the new copied one too.

* chore: ermoved actual make_public function

* refactor: handling return types to be uniform

---------

Co-authored-by: madclaws <pavanan.anandu63@gmail.com>