Custom certificates (#4064)
* add the create certificate support

* chore: crud apis wip

* checkout

* test: create-cert failure cases done

* feat: Added all CRUD apis and tests

* refactor: Added cert template type as enum

* refactor: removed dev logs

* feat: Added issued_certificate schema

* refactor: types

* fix: removed public validation, will add scope validation

* upload image in gcs

* test cases

* optimization

* fix test case

* Custom certificates: certificate template APIs (#4055)

* fix: gettext extract

* fix: Using enum for type in certificate_template schema + minor test changes

* feat: added functions for cert mgmt, tests to be added

* resolved reviews

* build: Run CIs for any PRs

* feat: Added new provider google_slides

We were using google_cloud_storage as provider for generating
certificates. But this has an issue, we cache the token. So if say
Bhasini call is ran first then the cache for gcs token will only have single
scope, so next time when we generate certificate, we get a token from cache
which had only single scope, instead of required scopes we need for generating
certificate.

* fix: Removed making copied template public

Since service account is copying the template, the service account will
have access to the new copied one too.

* checkout

* chore: ermoved actual make_public function

* feat: tracking cert issuing

* refactor: handling return types to be uniform

* fix: empty map check

* fix: notification type error

* fix: added unique remote name for generated certs + saving gcs_urls in issue_certificates

* fix: dialyzer errors

* fix: dialyzer errors

* feat: Added max_retries: 3 plug

* Custom certificate: webhook function to generate certificate  (#4056)

* add the create certificate support

* chore: crud apis wip

* checkout

* test: create-cert failure cases done

* feat: Added all CRUD apis and tests

* refactor: Added cert template type as enum

* refactor: removed dev logs

* fix: removed public validation, will add scope validation

* upload image in gcs

* test cases

* optimization

* fix test case

* resolved reviews

* build: Run CIs for any PRs

* feat: Added new provider google_slides

We were using google_cloud_storage as provider for generating
certificates. But this has an issue, we cache the token. So if say
Bhasini call is ran first then the cache for gcs token will only have single
scope, so next time when we generate certificate, we get a token from cache
which had only single scope, instead of required scopes we need for generating
certificate.

* fix: Removed making copied template public

Since service account is copying the template, the service account will
have access to the new copied one too.

* chore: ermoved actual make_public function

* refactor: handling return types to be uniform

---------

Co-authored-by: madclaws <pavanan.anandu63@gmail.com>

* fix: resolve conflicts

* fix: credo

* minor

* fix: More tests + refactoring

* Custom certificates: managing certificates issued to contacts (#4067)

* add the create certificate support

* chore: crud apis wip

* checkout

* test: create-cert failure cases done

* feat: Added all CRUD apis and tests

* refactor: Added cert template type as enum

* refactor: removed dev logs

* feat: Added issued_certificate schema

* refactor: types

* fix: removed public validation, will add scope validation

* upload image in gcs

* test cases

* optimization

* fix test case

* fix: Using enum for type in certificate_template schema + minor test changes

* feat: added functions for cert mgmt, tests to be added

* resolved reviews

* build: Run CIs for any PRs

* feat: Added new provider google_slides

We were using google_cloud_storage as provider for generating
certificates. But this has an issue, we cache the token. So if say
Bhasini call is ran first then the cache for gcs token will only have single
scope, so next time when we generate certificate, we get a token from cache
which had only single scope, instead of required scopes we need for generating
certificate.

* fix: Removed making copied template public

Since service account is copying the template, the service account will
have access to the new copied one too.

* checkout

* chore: ermoved actual make_public function

* feat: tracking cert issuing

* refactor: handling return types to be uniform

* fix: empty map check

* fix: notification type error

* fix: added unique remote name for generated certs + saving gcs_urls in issue_certificates

* fix: dialyzer errors

* fix: dialyzer errors

* feat: Added max_retries: 3 plug

* fix: resolve conflicts

* fix: credo

* minor

* fix: More tests + refactoring

* add validation in payload

* add tarams

* optimised code

* optimised code

* refactoring

* Custom Certificates: Added feature flag, metrics and tables to BQ  (#4072)

* wip: feature flag

* feat: Added feature-flags

* feat: Added certificate_issued metric

* refactor

---------

Co-authored-by: AmishaBisht <amishabisht33@gmail.com>

* fix: renamed feature flag

From custom_certificate_enabled to certificate_enabled for
making it consistent with the apis and FE

* Custom certificates: Deleting template copy and slide url validations (#4093)

* fix: fetch slide to verify the file

* fix: dialyzer

* fix: errors

* test: regression

* refactor

* refactor

* fix: tests

---------

Co-authored-by: madclaws <pavanan.anandu63@gmail.com>