fix: Will retry the last unsynced file, edge case
If the last media_file didnt sync first time, then for nxt 30 days if
there are no new media_file, then we were not syncing this file, because
of greater than only condition. This change makes sure that if max_id exists
and its even same as message_media_id, then we will queue it. If max_id is nil
that means no new media_file is there to sync