GCS: Media url sync enhancements (#4111)
* fix: message_media_id should not be older than 30 days

* fix: logging error to messages_media table

* fix: Added start from oldest unsynced file every nigth

* wip: tests, something fishy

* fix: logging err to message_media only in unsynced phase + tests

* wip: sending internal report

* feat: sending gcs media sync report

* fix: group_by instead of order_by + remove inspect

* fix: added COUNT to the query

* fix: tests + refactor + minor fixes

* refactor

* fix: queuingmsg_media which has gcs_url as nil

* fix: less than equal instead of less than

* test: removing flaky test which tests report gen

* refactor: removed unused publicist

* fix: using selected_as for using the alias in order_by unsyned files desc

* fix: Will retry the last unsynced file, edge case

If the last media_file didnt sync first time, then for nxt 30 days if
there are no new media_file, then we were not syncing this file, because
of greater than only condition. This change makes sure that if max_id exists
and its even same as message_media_id, then we will queue it. If max_id is nil
that means no new media_file is there to sync

* fix: added skip_organization_id in media_sync_data query

---------

Co-authored-by: Akhilesh Negi <akhileshnegi.an3@gmail.com>