fix: Return error if the flow is empty
For running a flow, we expect a start_node to be present
which will not be there for an empty flow. This was unhandled,
due to which high number of errors get raised if flow ran for
a collection with large number of contacts. Thus crossing the APM
request threshold.