fix: Handling filesearch-gpt kaapi timeout correctly
Previosuly if the message was nil on waking up a flow
If kaapi was enabled no matter what the case we were
fetching the latest webhook log and updating its error
and creating a "Failure" temp message. But this is errorneous
since even for a "no response" case in wait_for_response
we try to get the webhook_log and in the case of no webhook_log
found for a flow_context, we were getting error which crashes
the wakeup flow cron.

https://appsignal.com/project-tech4dev/sites/5f480c425ac13f7330101f30/exceptions/incidents/36